+++ src/@locale/app/lab/index.js
@@ -1,10 +1,12 @@
 import badges  from './badges'
 import buttons from './buttons'
+import i18n    from './i18n'
 
 const lab =
 {
     badges  ,
     buttons ,
+    i18n    ,
 };
 
 export default lab ;